chore: rename project to logging-with-context
Trying to dodge this error:

The name 'contextual-logging' is too similar to an existing project. See
https://pypi.org/help/#project-name for more information.
